requirements.txt
# Core dependencies
flask==2.0.1
flask_cors==4.0.0
werkzeug==2.3.7

# AI and ML
openai==0.28.0
google-cloud-aiplatform==1.36.0  # for vertexai
tiktoken==0.5.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2  # for embeddings

# Cloud Storage
google-cloud-storage==2.13.0
boto3==1.34.0  # AWS SDK

# Data Processing
pandas==2.1.4
opencv-python==4.8.1.78  # for cv2
numpy==1.24.3

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0